Move the customers query to CustomerHistory
