+++ plugins/woocommerce/src/Internal/Traits/OrderAttributionMeta.php
@@ -3,7 +3,6 @@ declare( strict_types=1 );
 
 namespace Automattic\WooCommerce\Internal\Traits;
 
-use Automattic\WooCommerce\Admin\API\Reports\Customers\Query as CustomersQuery;
 use Automattic\WooCommerce\Vendor\Detection\MobileDetect;
 use Exception;
 use WC_Meta_Data;
@@ -375,25 +374,5 @@ trait OrderAttributionMeta {
 		 * @param string $field_name  The field name.
 		 */
 		return (string) apply_filters( 'wc_order_attribution_field_description', $description, $field_name );
-	}
-
-	/**
-	 * Get the order history for the customer (data matches Customers report).
-	 *
-	 * @param int $customer_report_id The reports customer ID (not necessarily User ID).
-	 *
-	 * @return array|null Order count, total spend, and average spend per order.
-	 */
-	private function get_customer_history( $customer_report_id ): ?array {
-
-		$args = array(
-			'customers'   => array( $customer_report_id ),
-			// This defaults to a week ago for some reason.
-			'order_after' => null,
-		);
-
-		$customers_query = new CustomersQuery( $args );
-		$customer_data   = $customers_query->get_data();
-		return $customer_data->data[0] ?? null;
 	}
 }
